override devbox poetry config to enable scripts
Devbox's initHook.sh activates the poetry environment but uses the nix
installed python version.

The problem is that the scripts the python package installs are not
available to that python interpreter.

I'm not sure the is the _best_ fix but by activating the poetry
environment without overriding the python version we have access to the
installed scripts.

Perhaps it would be better to use the nix python but update a path
somewhere?